+++ json2csv_test.go
@@ -1,10 +1,8 @@
 package json2csv
 
 import (
-	"archive/zip"
 	"bytes"
 	"encoding/json"
-	"io"
 	"os"
 	"reflect"
 	"testing"
@@ -139,34 +137,10 @@ func TestJSON2CSV(t *testing.T) {
 	}
 }
 
-func NewJSONStreamZipReader(zipFileName string) JSONStreamReader {
-	zipReader, _ := zip.OpenReader(zipFileName)
-	return &JSONStreamZipReader{data: zipReader.File}
-}
-
-type JSONStreamZipReader struct {
-	data  []*zip.File
-	index int
-}
-
-func (jz *JSONStreamZipReader) HasNext() bool {
-	return jz.index < len(jz.data)
-}
-
-func (jz *JSONStreamZipReader) Read() map[string]interface{} {
-	child := jz.data[jz.index]
-	jz.index++
-	res := make(map[string]interface{})
-	cfd, _ := child.Open()
-	content, _ := io.ReadAll(cfd)
-	_ = json.Unmarshal(content, &res)
-	_ = cfd.Close()
-	return res
-}
-
 func TestJSON2CSVOnline(t *testing.T) {
+	zipFileName := "test.zip"
 	// extract csvHeader
-	reader := NewJSONStreamZipReader("test.zip")
+	reader := NewJSONStreamZipReader(zipFileName)
 	csvHeader, err := JSON2CSVHeader(reader)
 	if err != nil {
 		t.Errorf("Exception: %v", err)
@@ -174,11 +148,11 @@ func TestJSON2CSVOnline(t *testing.T) {
 	}
 
 	// extract row
-	output, err := os.Create("testFile.csv")
+	output, err := os.Create("test.csv")
 	if err != nil {
 		t.Errorf("Exception: %v", err)
 	}
-	reader = NewJSONStreamZipReader("test.zip")
+	reader = NewJSONStreamZipReader(zipFileName)
 	err = JSON2CSVOnline(reader, csvHeader, output)
 	if err != nil {
 		t.Errorf("ExceptionL %v", err)
